% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{crd_to_points}
\alias{crd_to_points}
\title{Coordinates to points}
\usage{
crd_to_points(lon, lat, CRS = 4326)
}
\arguments{
\item{lon}{longitude}

\item{lat}{latitude}

\item{CRS}{(optional) the EPSG code; uses wgs84 by default}
}
\value{
spatial features column
}
\description{
Generate sf points from lon/lat columns
}
\details{
Use this function to create sf points from lon and lat columns.
}
\examples{
data("pdx_breweries", package="spaceheater")
pts <- crd_to_points(pdx_breweries$lon, pdx_breweries$lat)
}
