% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{get_alphahull_polygon}
\alias{get_alphahull_polygon}
\title{Get alpha hull polygon from collection of points}
\usage{
get_alphahull_polygon(pts, tension = 0.5, buffer_size = 0)
}
\arguments{
\item{pts}{sf points}

\item{tension}{value between 0 and 1, controls how tight polygon should wrap around points;
value is fraction of diagonal of bounding box}

\item{buffer_size}{add an optional buffer around points; value in coordinate units}
}
\value{
sf GeometryCollection object with a polygon
}
\description{
A convenience wrapper around the alphahull::ashape() function.
The function creates a polygon around a collection of sf points, given inputs
for tension and an optional buffer.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x1 <- get_alphahull_polygon(pts)
x2 <- get_alphahull_polygon(pts, tension=0.2, buffer_size=0.1)
plot(pts)
plot(x1, add=TRUE)
plot(x2, add=TRUE)
}
