% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{get_epsg.default}
\alias{get_epsg.default}
\title{Get EPSG code for numeric}
\usage{
\method{get_epsg}{default}(x)
}
\arguments{
\item{x}{lonlat coordinates}
}
\value{
An integer with EPSG value
}
\description{
Get EPSG code for numeric
}
\seealso{
\code{\link{get_epsg()}}
}
