% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{get_epsg.sf}
\alias{get_epsg.sf}
\title{Get EPSG code for SF Object}
\usage{
\method{get_epsg}{sf}(x)
}
\arguments{
\item{x}{sf object}
}
\value{
An integer with EPSG value
}
\description{
Get EPSG code for SF Object
}
\seealso{
\code{\link{get_epsg()}}
}
