% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{get_epsg.sfc}
\alias{get_epsg.sfc}
\title{Get EPSG code for SF Object}
\usage{
\method{get_epsg}{sfc}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Get EPSG code for SF Object
}
\seealso{
\code{\link{get_epsg()}}
}
