% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{outliers_by_box}
\alias{outliers_by_box}
\title{Spatial outliers: BOX Method}
\usage{
outliers_by_box(pts, pct = NA)
}
\arguments{
\item{pts}{sf points}

\item{pct}{(optional) quantile to determine distance from center}
}
\value{
list with outlier status and sf polygon
}
\description{
Spatial outlier method that uses distance from a central point to draw a box
around a collection of points.  The distance to box boundaries is determined from the 90th
percentile, or by user input.
TODO: add manual center-point option, different xy distances
TODO: adjust code for new epsg function
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- outliers_by_box(pts)
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=TRUE)
plot(x[[2]], add=TRUE)
}
\seealso{
Other outlier functions: 
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_cluster}()},
\code{\link{outliers_by_density}()},
\code{\link{outliers_by_iforest}()},
\code{\link{outliers_by_polygon}()},
\code{\link{spatial_outliers}()}
}
\concept{outlier functions}
