% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{outliers_by_cluster}
\alias{outliers_by_cluster}
\title{Spatial outliers: CLUSTER Method}
\usage{
outliers_by_cluster(
  pts,
  eps = 0.25,
  MinPts = 5,
  tension = 0.5,
  buffer_size = 0.05
)
}
\arguments{
\item{pts}{sf points}

\item{eps}{parameter passed to dbscan, distance between neighboring points}

\item{MinPts}{parameter passed to dbscan, min points per cluster}

\item{tension}{(optional) value from 0 to 1 to control how tight polygon should wrap around points}

\item{buffer_size}{(optional) buffer around exterior points, in coordinate units}
}
\value{
list with outlier status and sf polygon
}
\description{
Spatial outlier method based on the dbscan spatial clustering algorithm.  User should
run dbscan::kNNdistplot() first to get a reasonable value for eps.  The dbscan algorithm
clusters according to a nominal distance between points (eps) and minimum number of
points per cluster (MinPts) and puts all points that don't neatly fit into a group
into cluster zero.  This method draws a polygon around the non-cluster-zero points,
applies some edge corrections, and determines outlier status accordingly.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- outliers_by_cluster(pts)
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=TRUE)
plot(x[[2]], add=TRUE)
}
\seealso{
\code{\link{get_alphahull_polygon()}}

Other outlier functions: 
\code{\link{outliers_by_box}()},
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_density}()},
\code{\link{outliers_by_iforest}()},
\code{\link{outliers_by_polygon}()},
\code{\link{spatial_outliers}()}
}
\concept{outlier functions}
