% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{outliers_by_density}
\alias{outliers_by_density}
\title{Spatial outliers: DENSITY Method}
\usage{
outliers_by_density(pts, thresh = 0.1, sigma = 0.1)
}
\arguments{
\item{pts}{sf points}

\item{thresh}{fraction of maximum density to use as outlier threshold}

\item{sigma}{value passed to spatstat for kernel density size;
the value is used as standard deviation of Gaussian for smoothing}
}
\value{
list with outlier status and sf polygon
}
\description{
Spatial outlier method that uses kernel density to determine outlier status.
The algorithm uses spatstat to compute a density grid, grDevices to create
contour lines based on user-defined threshold value, and then determines outliers
by intersecting points with the contours.  The threshold value is based on some
fraction of the maximum density, 0.1 in the absence of user input.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- outliers_by_density(pts)
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=T)
plot(x[[2]], add=T)
}
\seealso{
\code{\link{points_to_density()}}

Other outlier functions: 
\code{\link{outliers_by_box}()},
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_cluster}()},
\code{\link{outliers_by_iforest}()},
\code{\link{outliers_by_polygon}()},
\code{\link{spatial_outliers}()}
}
\concept{outlier functions}
