% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{outliers_by_iforest}
\alias{outliers_by_iforest}
\title{Spatial outliers: IFOREST Method}
\usage{
outliers_by_iforest(pts, thresh = 0.95, tension = 0.3, buffer_size = 0.05)
}
\arguments{
\item{pts}{sf points}

\item{thresh}{percentile value to use for determining outlier status}

\item{tension}{(optional) value from 0 to 1 to control how tight polygon should wrap around points}

\item{buffer_size}{(optional) buffer around exterior points, in coordinate units}
}
\value{
list with outlier status and sf polygon
}
\description{
Spatial outliers based on the Isolation Forest algorithm implemented in the
solitude package.  IForest is a tree-based method used to determine nominal difference
from normal values; it is not traditionally applied to spatial data, so this
method is really just experimental.  Threshold value defines the percentile to
use for determining outlier status.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- outliers_by_iforest(pts)
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=TRUE)
plot(x[[2]], add=TRUE)
}
\seealso{
Other outlier functions: 
\code{\link{outliers_by_box}()},
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_cluster}()},
\code{\link{outliers_by_density}()},
\code{\link{outliers_by_polygon}()},
\code{\link{spatial_outliers}()}
}
\concept{outlier functions}
