% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{outliers_by_polygon}
\alias{outliers_by_polygon}
\title{Spatial outliers: POLYGON Method}
\usage{
outliers_by_polygon(
  pts,
  center_point = NA,
  pct = NA,
  tension = 0.3,
  buffer_size = 0.05
)
}
\arguments{
\item{pts}{sf points}

\item{center_point}{(optional) sf point with location to measure from}

\item{pct}{(optional) percentile to use for determining distance from center point}

\item{tension}{(optional) value from 0 to 1 to control how tight polygon should wrap around points}

\item{buffer_size}{(optional) buffer around exterior points, in coordinate units}
}
\value{
list with outlier status and sf polygon
}
\description{
Spatial outlier method that uses distance from a central point to draw a bounding polygon
from which to determine outlier status.  The outliers should be identical to the circle
method, except the shape will account for anisotropy in the the point pattern.  The polygon
is built with alpha hull, so the user can specify tension and buffer size.  Note that, for
very low (tighter) tension values edge effects may lead to differences from the circle method.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- outliers_by_polygon(pts)
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=T)
plot(x[[2]], add=T)
}
\seealso{
\code{\link{get_alphahull_polygon}}

Other outlier functions: 
\code{\link{outliers_by_box}()},
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_cluster}()},
\code{\link{outliers_by_density}()},
\code{\link{outliers_by_iforest}()},
\code{\link{spatial_outliers}()}
}
\concept{outlier functions}
