% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdx_breweries.R
\docType{data}
\name{pdx_breweries}
\alias{pdx_breweries}
\title{Breweries near Portland, OR}
\format{
A data frame woth 251 rows and 6 variables:
\describe{
\item{byname}{brewery name}
\item{lon}{longitude}
\item{lat}{latitude}
\item{type}{brewery type}
\item{geometry}{coordinates}
\item{dist_Portland}{distance from Portland, OR, in km}
}
}
\source{
derived from untappd.com in 2021, geocoded with a variety of tools
}
\usage{
pdx_breweries
}
\description{
Dataset containing names and locations of breweries within 150 km of
Portland, OR.
}
\keyword{datasets}
