% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceheater_package.R
\docType{package}
\name{spaceheater-package}
\alias{spaceheater-package}
\title{Spatial Analysis of Satellite and Raster Data in R}
\description{
Spatial Analysis of Satellite and Raster Data in R
}
\author{
Neal Thomas Barsch
\cr
\cr
\emph{Maintainer:} Neal Thomas Barsch \email{neal.barsch@gmail.com}
}
\keyword{package}
