% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{spaceheater_epsg}
\alias{spaceheater_epsg}
\title{Get EPSG Code for local UTM projection}
\usage{
spaceheater_epsg(lon, lat)
}
\arguments{
\item{lon}{longitude GPS coordinate}

\item{lat}{latitude GPS coordinate}
}
\value{
integer value for EPSG code
}
\description{
Compute the EPSG code for local UTM projection, based on GPS inputs.
Calculation is based on standard formula:
32700 - round((45+lat)/90)*100 + round((183+lon)/6)
}
\examples{
crd <- c(-76.3451, 34.3352)
spaceheater_epsg(crd[1], crd[2])
}
