% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceout.R
\name{spatial_outliers}
\alias{spatial_outliers}
\title{Spatial outliers}
\usage{
spatial_outliers(pts, method = "polygon", ...)
}
\arguments{
\item{pts}{sf points}

\item{method}{which method to use, "box", "circle", "polygon", "density", "cluster", "iforest"}

\item{...}{input parameters specific to individual methods}
}
\value{
list with outlier status and sf polygon
}
\description{
A convenience wrapper used to call any of the spatial outlier methods implemented
in \code{spaceheater}.
}
\examples{
data("pdx_breweries")
pts <- pdx_breweries$geometry
x <- spatial_outliers(pts,method="polygon")
plot(pts)
plot(pts[x[[1]]==-1], col="red", add=TRUE)
plot(x[[2]], add=TRUE)
}
\seealso{
Other outlier functions: 
\code{\link{outliers_by_box}()},
\code{\link{outliers_by_circle}()},
\code{\link{outliers_by_cluster}()},
\code{\link{outliers_by_density}()},
\code{\link{outliers_by_iforest}()},
\code{\link{outliers_by_polygon}()}
}
\concept{outlier functions}
