% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spaceutils.R
\name{get_epsg}
\alias{get_epsg}
\title{Get EPSG Code}
\usage{
get_epsg(x)
}
\arguments{
\item{x}{Coordinates or spatial object}
}
\value{
An integer with EPSG value
}
\description{
Get EPSG code for local UTM projection
}
\details{
This is a generic method to determine the EPSG code for a set of coordinates
\code{epsg.numeric} will take a lonlat vector whereas \code{epsg.sfc} takes an
sfc object and determines the EPSG code from the center of the bounding box.
Computation is based on a method outlined in the text Geocomputation in R,
by Robin Lovelace, section 7.3
}
\examples{
lonlat <- c(-76.2, 33.5)
pt <- sf::st_point(lonlat)

get_epsg(lonlat)
get_epsg(pt)
}
